@@ -1,7 +1,7 @@
 package com.processout.processout_sdk;
 
 import android.app.Activity;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 public class TestActivity extends Activity {
